feat(ui): expose per-repo automation settings in Backlog (no task needed)
Extract the in-task Automation popover body into a shared, layout-neutral
AutomationSettings.svelte (showHeader + armCoachmarks props, coachIf wrapper
to avoid coachTarget id collisions across two mounts). AutomationPanel becomes
a thin popover shell wrapping it.

Add an Automation tab to the Backlog drill-down (both tab bars) rendering the
full per-repo settings for the selected repo, scoped to a height-owning
scroller. Thread store.drain through all three BacklogView mounts (direct
mobile/desktop + BacklogOverlay) so the epic banner + drain-cap are live.

Relax buildBacklogPayload to also list forge-less repos as kind:"local" so
every repo's settings are reachable without launching a task (lightweight
repos already appeared via LocalForge).
