+++ ui/messages/de.json
@@ -1577,5 +1577,8 @@
   "diagnostics_hint_gitcap_ok": "git unterstützt Lightweight-Merge (≥ 2.38).",
   "diagnostics_hint_gitcap_old": "git ≥ 2.38 nicht verfügbar — der Lightweight-Repo-Modus benötigt git merge-tree --write-tree (git ≥ 2.38) für lokale Merges. git fehlt möglicherweise (siehe git-Prüfung) oder ist zu alt; installiere oder aktualisiere es, um Lightweight-Repos zu nutzen.",
   "diagnostics_hint_gh_not_required": "GitHub CLI ist nicht installiert oder nicht authentifiziert, es sind jedoch keine Forge-basierten Repos konfiguriert — gh ist optional, wenn ausschließlich Lightweight-Repos verwendet werden.",
-  "diagnostics_label_git_mergetree": "git merge-tree"
+  "diagnostics_label_git_mergetree": "git merge-tree",
+  "backlog_tab_automation": "Automatisierung",
+  "feat_repo_settings_backlog_title": "Repo-Einstellungen ohne Task",
+  "feat_repo_settings_backlog_body": "Die Repo-Automatisierungseinstellungen — Kritiker, Plan-Gate, Autopilot, Auto-Drain/Merge, Sandbox-Profil, Standardmodell, Drain-Konfiguration sowie Reviewer-/Merger-Rollen — sind jetzt als Automatisierungs-Tab in der Backlog-Detailansicht verfügbar. Stelle die Vorgaben jedes Repos ein, ohne einen Task zu starten; Änderungen teilen sich eine einzige Quelle der Wahrheit mit dem Task-Panel."
 }
