+++ ui/messages/en.json
@@ -1577,5 +1577,8 @@
   "diagnostics_hint_gitcap_ok": "git supports lightweight merge (≥ 2.38).",
   "diagnostics_hint_gitcap_old": "git ≥ 2.38 not available — lightweight repo mode needs git merge-tree --write-tree (git ≥ 2.38) for local merges. git may be missing (see the git check) or too old; install or update it to use lightweight repos.",
   "diagnostics_hint_gh_not_required": "GitHub CLI is not installed or not authenticated, but no forge-backed repos are configured — gh is optional when using only lightweight repos.",
-  "diagnostics_label_git_mergetree": "git merge-tree"
+  "diagnostics_label_git_mergetree": "git merge-tree",
+  "backlog_tab_automation": "Automation",
+  "feat_repo_settings_backlog_title": "Repo settings without a task",
+  "feat_repo_settings_backlog_body": "The per-repo automation settings — critic, plan gate, autopilot, auto-drain/merge, sandbox profile, default model, drain config and reviewer/merger roles — are now an Automation tab in the Backlog drill-down. Tune any repo's defaults without launching a task; changes share one source of truth with the in-task panel."
 }
