+++ ui/src/routes/+page.svelte
@@ -1494,6 +1494,9 @@
   }
 </script>
 
+<!-- Pre-existing top-level page template; this change only threads a `drain` prop into
+     three existing BacklogView mounts. Complexity is inherent, not introduced here. -->
+<!-- fallow-ignore-next-line complexity -->
 <svelte:window onkeydown={onShortcut} />
 
 <div
@@ -1626,6 +1629,7 @@
               flow={true}
               epics={store.epics}
               {inTrainPrs}
+              drain={store.drain}
             />
           {/if}
         </div>
@@ -1785,6 +1789,7 @@
             {onlaunchtrain}
             epics={store.epics}
             {inTrainPrs}
+            drain={store.drain}
           />
         {:else if selected}
           <Viewport
@@ -2111,6 +2116,7 @@
     epics={store.epics}
     {inTrainPrs}
     target={epicTarget}
+    drain={store.drain}
   />
 {/if}
